% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BBOX_SSA}
\alias{BBOX_SSA}
\title{Study-area bounding box around SIMCOSTA buoy 515 / Salvador, Bahia (WGS84)
Approx 16 km x 22 km covering Todos os Santos Bay entrance}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
BBOX_SSA
}
\description{
Study-area bounding box around SIMCOSTA buoy 515 / Salvador, Bahia (WGS84)
Approx 16 km x 22 km covering Todos os Santos Bay entrance
}
\keyword{datasets}
