% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib-processing.R
\name{crop_grib_to_extent}
\alias{crop_grib_to_extent}
\title{Crop GRIB layers to study extent}
\usage{
crop_grib_to_extent(u_layers, v_layers, study_area, records_sf)
}
\arguments{
\item{u_layers, v_layers}{SpatRaster U and V components (WGS84).}

\item{study_area}{\code{sfc} polygon defining the study area (any CRS).}

\item{records_sf}{\code{sf} POINT object of GPS records (any CRS).}
}
\value{
A list with elements \code{u_study} and \code{v_study} (cropped SpatRasters).
}
\description{
Computes the combined bounding box of a study area polygon and GPS records,
then crops U and V layers to that extent.
}
